Unstub ballot measure "total_contributions"
In the /referendum/:id/supporting and /referendum/:id/opposing
endpoints, this commit unstubs the "total_contributions" value.

This re-uses the referendum_supporters_calculator since the data is
already there, although it is a bit weird and this calculator is going
to get super complicated. But we can deal with that later.

[Fixes #23]