Output all Oakland Candidates
The frontend needs the links to all the committee pages to exist, and
the candidate spreadsheet was out of date from last year.

There is a new candidate spreadsheet (the "All Oakland Candidates" tab)
that appears to be a raw dump from netfile. Although many of those
committees won't be relevant, let's use it anyway just to be safe.

Some items on that spreadsheet have "Pending" Filer_IDs, so I've added
defensive logic.